requirements.txt
python-telegram-bot==20.0
python-dotenv==0.21.0
google-genai==0.1.0
requests==2.28.1
streamlit==1.12.0
streamlit-webrtc==0.45.0
altair==4.2.0
av==9.2.0
